+++ test/ResourceGenerator/UrlBasedCollectionWithRouteParamsTest.php
@@ -49,8 +49,8 @@ class UrlBasedCollectionWithRouteParamsTest extends TestCase
             'foo-bar',
             self::getObjectPropertyHydratorClass(),
             'id',
-            'bar_id',
-            ['foo_id' => 1234]
+            ['foo_id' => 1234],
+            ['id' => 'bar_id']
         );
 
         $metadataMap->has(TestAsset\FooBar::class)->willReturn(true);
@@ -121,8 +121,8 @@ class UrlBasedCollectionWithRouteParamsTest extends TestCase
             'foo-bar',
             self::getObjectPropertyHydratorClass(),
             'id',
-            'bar_id',
-            ['foo_id' => 1234]
+            ['foo_id' => 1234],
+            ['id' => 'bar_id']
         );
 
         $metadataMap->has(TestAsset\FooBar::class)->willReturn(true);